+++ package.json
@@ -16,7 +16,8 @@
     "ink": "^5.0.1",
     "is-wsl": "^3.1.0",
     "open": "^10.1.0",
-    "react": "^18.3.1"
+    "react": "^18.3.1",
+    "terminal-link": "^3.0.0"
   },
   "devDependencies": {
     "@oclif/plugin-command-snapshot": "^5.2.3",
